Connect PC comms to floodlight, btn and knob code.
