+++ Core/Src/main.c
@@ -24,6 +24,8 @@
 #include "math.h"
 #include <stdlib.h>
 #include "app_usart.h"
+#include "floodlight.h"
+#include "ui.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -54,72 +56,6 @@ DMA_HandleTypeDef hdma_usart2_rx;
 DMA_HandleTypeDef hdma_usart2_tx;
 
 /* USER CODE BEGIN PV */
-typedef enum {
-	//NOTE: repeat since latter half knobs are on adc2
-	KNOB1Chn = ADC_CHANNEL_1, KNOB2Chn = ADC_CHANNEL_2, KNOB3Chn = ADC_CHANNEL_17,
-	KNOB4Chn = ADC_CHANNEL_13,	KNOB5Chn = ADC_CHANNEL_3, KNOB6Chn = ADC_CHANNEL_4
-} KnobChannel;
-
-struct FloodlightLED {
-	GPIO_TypeDef * GPIO;
-	uint32_t LEDColourChannelPin;
-	_Bool pinOn;
-	unsigned short dim;
-	unsigned short rise1;
-	unsigned short fall1;
-	unsigned short rise2;
-	unsigned short fall2;
-};
-
-struct Knob {
-	ADC_HandleTypeDef * adc;
-	KnobChannel channel;
-	uint16_t value;
-};
-
-struct Button {
-	GPIO_TypeDef * GPIO;
-	uint32_t pin;
-	GPIO_PinState pinState;
-	_Bool buttonState;
-	uint8_t debounceCounter;
-};
-
-//2^32 - 100000
-#define DWT_CYCLE_RESET 4294867296
-
-#define NUMLIGHTS 2
-#define COLOURCHANNELSPERLIGHT 3
-struct FloodlightLED floodlights[NUMLIGHTS][COLOURCHANNELSPERLIGHT];
-
-#define NUMKNOBS 6
-//Div by 5 since 5ms polling
-#define DEBOUNCECYCLES 20
-struct Knob knobs[NUMKNOBS];
-
-#define NUMBUTTONS 6
-struct Button buttons[NUMBUTTONS];
-
-_Bool initialized = 0;
-
-unsigned long width = 0;
-unsigned long avgWidth = 0;
-_Bool set = 0;
-unsigned long start = 0;
-unsigned short place = 0;
-
-unsigned short curMains = 0;
-uint8_t incrDuty = 0;
-int curShift = 0;
-uint32_t curStartTilNow = 0;
-
-volatile uint32_t *DWT_CONTROL = (uint32_t*) 0xE0001000;
-volatile uint32_t *DWT_CYCCNT = (uint32_t*) 0xE0001004;
-volatile uint32_t *DEMCR = (uint32_t*) 0xE000EDFC;
-volatile uint32_t *LAR = (uint32_t*) 0xE0001FB0; // <-- added lock access register
-
-const uint32_t startReset = (uint32_t) (pow(2, sizeof(uint32_t) * 8));
-const uint32_t startTilNowReset = (uint32_t) (pow(2, sizeof(uint32_t) * 8) / 64);
 
 /* USER CODE END PV */
 
@@ -134,371 +70,30 @@ static void MX_TIM4_Init(void);
 static void MX_TIM3_Init(void);
 static void MX_USART2_UART_Init(void);
 /* USER CODE BEGIN PFP */
-_Bool initializeFloodlightStructs(void);
-_Bool initializeKnobStructs(void);
-_Bool initializeButtonStructs(void);
-uint16_t readADCChannel(KnobChannel channel, ADC_HandleTypeDef * adc);
-void pollKnobs(void);
-void debounceButtons(void);
-void resetRiseFall(uint8_t floodlightNum, uint8_t LEDNum, _Bool *dimCorrect);
-void mainsDetect(void);
-void pulseFloodlight(void);
-void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin);
-void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim);
+
 /* USER CODE END PFP */
 
 /* Private user code ---------------------------------------------------------*/
 /* USER CODE BEGIN 0 */
-//Set to (666+1)*(7+1) so same rate as timer2
-#define timer2cycle() (*DWT_CYCCNT/5336)
-#define starttilnow() (timer2cycle() - start)
 
-#define ADC_WIDTH 4092
-#define HALF_ADC_WIDTH 2048
+volatile uint32_t *DWT_CONTROL = (uint32_t*) 0xE0001000;
+volatile uint32_t *DWT_CYCCNT = (uint32_t*) 0xE0001004;
+volatile uint32_t *DEMCR = (uint32_t*) 0xE000EDFC;
+volatile uint32_t *LAR = (uint32_t*) 0xE0001FB0; // <-- added lock access register
 
-#define MAX_LED_DIM 255
-#define NUM_LED_LEVELS 256 //255+1
-
-#define LED_ON_STATE(floodlightNum, LEDNum) if(!floodlights[floodlightNum][LEDNum].pinOn) {\
-		HAL_GPIO_WritePin(floodlights[floodlightNum][LEDNum].GPIO, \
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin, GPIO_PIN_SET);\
-		floodlights[floodlightNum][LEDNum].pinOn = 1;}
-#define LED_OFF_STATE(floodlightNum, LEDNum) if(floodlights[floodlightNum][LEDNum].pinOn) {\
-		HAL_GPIO_WritePin(floodlights[floodlightNum][LEDNum].GPIO, \
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin, GPIO_PIN_RESET);\
-		floodlights[floodlightNum][LEDNum].pinOn = 0;}
-
-
-_Bool initializeFloodlightStructs() {
-	short floodlightNum, LEDNum;
-	for (floodlightNum = 0; floodlightNum < NUMLIGHTS;
-			floodlightNum++) {
-		for (LEDNum = 0; LEDNum < COLOURCHANNELSPERLIGHT; LEDNum++) {
-			//NOTE: assuming no more than 10 LED per floodlight
-			switch (10*floodlightNum + LEDNum) {
-			case 0:
-				floodlights[floodlightNum][LEDNum].GPIO = GPIOB;
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin = Fld1R_Pin;
-				break;
-			case 1:
-				floodlights[floodlightNum][LEDNum].GPIO = GPIOF;
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin = Fld1G_Pin;
-				break;
-			case 2:
-				floodlights[floodlightNum][LEDNum].GPIO = GPIOF;
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin = Fld1B_Pin;
-				break;
-			case 10:
-				floodlights[floodlightNum][LEDNum].GPIO = GPIOA;
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin = Fld2R_Pin;
-				break;
-			case 11:
-				floodlights[floodlightNum][LEDNum].GPIO = GPIOA;
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin = Fld2G_Pin;
-				break;
-			case 12:
-				floodlights[floodlightNum][LEDNum].GPIO = GPIOB;
-				floodlights[floodlightNum][LEDNum].LEDColourChannelPin = Fld2B_Pin;
-				break;
-			default:
-				//Configure shit!
-				return 1;
-			}
-			floodlights[floodlightNum][LEDNum].pinOn = 0;
-			floodlights[floodlightNum][LEDNum].dim = 0;
-			floodlights[floodlightNum][LEDNum].rise1 = 0;
-			floodlights[floodlightNum][LEDNum].fall1 = 0;
-			floodlights[floodlightNum][LEDNum].rise2 = 0;
-			floodlights[floodlightNum][LEDNum].fall2 = 0;
-
-		}
-	}
-	return 0;
-}
-
-_Bool initializeKnobStructs() {
-	uint8_t knob;
-	for(knob = 0; knob < NUMKNOBS; knob++) {
-		switch(knob) {
-		case 0:
-			knobs[knob].channel = KNOB1Chn;
-			knobs[knob].adc =&hadc1;
-			break;
-		case 1:
-			knobs[knob].channel = KNOB2Chn;
-			knobs[knob].adc = &hadc1;
-			break;
-		case 2:
-			knobs[knob].channel = KNOB3Chn;
-			knobs[knob].adc = &hadc2;
-			break;
-		case 3:
-			knobs[knob].channel = KNOB4Chn;
-			knobs[knob].adc = &hadc2;
-			break;
-		case 4:
-			knobs[knob].channel = KNOB5Chn;
-			knobs[knob].adc = &hadc2;
-			break;
-		case 5:
-			knobs[knob].channel = KNOB6Chn;
-			knobs[knob].adc = &hadc2;
-			break;
-		default:
-			//Configure shit!
-			return 1;
-		}
-
-		knobs[knob].value = 0;
-	}
-	return 0;
-}
-
-
-_Bool initializeButtonStructs() {
-	uint8_t button;
-	for(button = 0; button < NUMBUTTONS; button++) {
-		switch(button) {
-		case 0:
-			buttons[button].GPIO = GPIOA;
-			buttons[button].pin = But1_Pin;
-			break;
-		case 1:
-			buttons[button].GPIO = GPIOA;
-			buttons[button].pin = But2_Pin;
-			break;
-		case 2:
-			buttons[button].GPIO = GPIOA;
-			buttons[button].pin = But3_Pin;
-			break;
-		case 3:
-			buttons[button].GPIO = GPIOB;
-			buttons[button].pin = But4_Pin;
-			break;
-		case 4:
-			buttons[button].GPIO = GPIOB;
-			buttons[button].pin = But5_Pin;
-			break;
-		case 5:
-			buttons[button].GPIO = GPIOA;
-			buttons[button].pin = But6_Pin;
-			break;
-		default:
-			//Configure shit!
-			return 1;
-		}
-
-		buttons[button].pinState = GPIO_PIN_RESET;
-		buttons[button].buttonState = 0;
-		buttons[button].debounceCounter = 0;
-	}
-	return 0;
-}
-
-uint16_t readADCChannel(KnobChannel channel, ADC_HandleTypeDef * adc) {
-	ADC_ChannelConfTypeDef sConfig = { 0 };
-
-	/** Configure Regular Channel
-	 */
-	sConfig.Channel = channel;
-	sConfig.Rank = ADC_REGULAR_RANK_1;
-	sConfig.SingleDiff = ADC_SINGLE_ENDED;
-	sConfig.SamplingTime = ADC_SAMPLETIME_2CYCLES_5;
-	sConfig.OffsetNumber = ADC_OFFSET_NONE;
-	sConfig.Offset = 0;
-	if (HAL_ADC_ConfigChannel(&hadc1, &sConfig) != HAL_OK) {
-		Error_Handler();
-	}
-
-	HAL_ADC_Start(adc);
-	HAL_ADC_PollForConversion(adc, 1);
-	return HAL_ADC_GetValue(adc);
-}
-
-void pollKnobs() {
-	static uint8_t knob;
-	for(knob = 0; knob < NUMKNOBS; knob++) {
-		//Smooth to filter out noise
-		knobs[knob].value = (9*knobs[knob].value +
-				readADCChannel(knobs[knob].channel, knobs[knob].adc)) / 10;
-	}
-}
-
-void debounceButtons() {
-	static uint8_t button;
-	//TODO: get rid of this
-	static int temp = 0;
-
-	for (button = 0; button < NUMBUTTONS; button++) {
-		if (HAL_GPIO_ReadPin(buttons[button].GPIO, buttons[button].pin)
-				== GPIO_PIN_SET) {
-			if (buttons[button].buttonState == 0) {
-				buttons[button].debounceCounter++;
-				if (buttons[button].debounceCounter >= DEBOUNCECYCLES) {
-					//Legit button press!
-					temp = 1;
-					buttons[button].debounceCounter = 0;
-					buttons[button].buttonState = 1;
-				}
-			}
-		} else if (buttons[button].debounceCounter != 0) {
-			buttons[button].debounceCounter = 0;
-		} else if (buttons[button].buttonState != 0) {
-			buttons[button].buttonState = 0;
-		}
-	}
-	if (temp == 1) {
-				temp = 0;
-	}
-}
-
-void resetRiseFall(uint8_t floodlightNum, uint8_t LEDNum, _Bool *dimCorrect) {
-	//Divide by 2 since we cycle duty per half-wave, not full wave
-		static unsigned short effDuty;
-	effDuty = (short) ((((unsigned long) width
-			* (unsigned long) floodlights[floodlightNum][LEDNum].dim)) / (2 * NUM_LED_LEVELS));
-	//If dim is odd, this granularity will be lost in compression since only 128 pulse width levels available
-	//i.e. (512*255)/(2*256) == 255 == (512*254)/(2*256)
-	//Circumvent this by offsetting all odd values by 1 on fall to increase pulse width by 1/2 a level
-	//NOTE passing in as pointer since resetRiseFall can be triggered asynchronously via seperate interrupts
-	*dimCorrect = floodlights[floodlightNum][LEDNum].dim & 1;
-	//TODO: calculate & set timer freq so exactly 512*mains freq (change if using ext osc)
-	floodlights[floodlightNum][LEDNum].rise1 = ((width / 2) - effDuty) / 2;
-	floodlights[floodlightNum][LEDNum].fall1 = floodlights[floodlightNum][LEDNum].rise1 + effDuty + *dimCorrect;
-	floodlights[floodlightNum][LEDNum].rise2 = (width / 2) + floodlights[floodlightNum][LEDNum].rise1;
-	floodlights[floodlightNum][LEDNum].fall2 = (width / 2) + floodlights[floodlightNum][LEDNum].fall1;
-}
-void resetAllRiseFall(_Bool *dimCorrect) {
-	static short floodlightNum, LEDNum;
-		for (floodlightNum = 0; floodlightNum < NUMLIGHTS;
-				floodlightNum++) {
-			for (LEDNum = 0; LEDNum < COLOURCHANNELSPERLIGHT; LEDNum++) {
-				resetRiseFall(floodlightNum, LEDNum, dimCorrect);
-			}
-		}
-}
-
-void mainsDetect() {
-	//Why does width occasionally jump to ~20k micros for a bit??  Then returns back to ~61Hz
-	//Filtered out by blip discard method.  May be artifact of low frequency operation
-	if (set) {
-		curStartTilNow = starttilnow();
-		//Slowly track to actual width if permanent shift
-		avgWidth = ((19 * (unsigned long) width + (curStartTilNow)) / 20);
-		//Try to filter out blips, is this an actual frequency shift on mains?
-		//If average closer to calculated than historical, this is the new frequency
-		//Explicit cast to (long) since stdlib has no overload of abs for unsigned long
-      if (width == 0 || abs((unsigned long)((curStartTilNow) - width)) < width/16
-       || abs((unsigned long)((curStartTilNow) - avgWidth)) < abs((unsigned long)(width - avgWidth))) {
-        width = curStartTilNow;
-		//Reset avg
-		//avgWidth = width;
-		//reset applicable variables
-        static _Bool dimCorrect;
-		resetAllRiseFall(&dimCorrect);
-      }
-
-	} else {
-		set = 1;
-	}
-	if (*DWT_CYCCNT > DWT_CYCLE_RESET)  *DWT_CYCCNT = 0;                  // clear DWT cycle counter
-	start = timer2cycle();
-
-
-	//--------------------------------------------------------------------------------------------------------
-	//ONLY FOR TESTING
-	static _Bool flip1, flip2;
-	static uint16_t reset = 0;
-		reset++;
-		if (reset >= 1) {
-			if (floodlights[0][0].dim > MAX_LED_DIM) floodlights[0][0].dim=0;
-			if (floodlights[0][1].dim > MAX_LED_DIM) floodlights[0][1].dim=0;
-			if(!flip1) {
-			floodlights[0][0].dim = (floodlights[0][0].dim + 1);
-			if (floodlights[0][0].dim >= MAX_LED_DIM-1) {
-				flip1 = 1;
-				floodlights[0][0].dim = MAX_LED_DIM-1;
-			}
-			} else {
-				floodlights[0][0].dim = (floodlights[0][0].dim - 1);
-				if (floodlights[0][0].dim <= 0 || floodlights[0][0].dim > MAX_LED_DIM) {
-					flip1 = 0;
-					floodlights[0][0].dim = 0;
-				}
-			}
-			if(!flip2) {
-			floodlights[0][1].dim = (floodlights[0][1].dim - 1);
-			if (floodlights[0][1].dim == MAX_LED_DIM-1) flip2 = 1;
-			if (floodlights[0][1].dim <= 0 || floodlights[0][1].dim > MAX_LED_DIM ) {
-				flip2 = 1;
-				floodlights[0][1].dim = 0;
-			}
-			} else {
-				floodlights[0][1].dim = (floodlights[0][1].dim + 1);
-				if (floodlights[0][0].dim >= MAX_LED_DIM-1) {
-					flip2 = 1;
-					floodlights[0][1].dim = MAX_LED_DIM-1;
-				}
-			}
-					floodlights[0][2].dim = abs(floodlights[0][0].dim - floodlights[0][1].dim) % MAX_LED_DIM;
-			        static _Bool dimCorrect;
-					resetRiseFall(0, 0, &dimCorrect);
-					resetRiseFall(0, 1, &dimCorrect);
-					resetRiseFall(0, 2, &dimCorrect);
-					reset = 0;
-					floodlights[1][0].dim = 255 - floodlights[0][0].dim;
-					floodlights[1][1].dim = 255 - floodlights[0][1].dim;
-					floodlights[1][2].dim = 255 - floodlights[0][2].dim;
-		}
-		//--------------------------------------------------------------------------------------------------------
-}
-
-#pragma GCC push_options
-#pragma GCC optimize ("-O3")
-void pulseFloodlight() {
-	//Triggers every 32 us
-	place = starttilnow();
-	static short floodlightNum, LEDNum;
-	for (floodlightNum = 0; floodlightNum < NUMLIGHTS;
-			floodlightNum++) {
-		for (LEDNum = 0; LEDNum < COLOURCHANNELSPERLIGHT; LEDNum++) {
-			//NOTE: we know as each success LT is evaluated, don't need to re-eval subsequent GTE
-			if (place < floodlights[floodlightNum][LEDNum].rise1) {
-				LED_OFF_STATE(floodlightNum, LEDNum)
-			} else if (place < floodlights[floodlightNum][LEDNum].fall1) {
-				LED_ON_STATE(floodlightNum, LEDNum)
-			} else if (place < floodlights[floodlightNum][LEDNum].rise2) {
-				LED_OFF_STATE(floodlightNum, LEDNum)
-			} else if (place < floodlights[floodlightNum][LEDNum].fall2) {
-				LED_ON_STATE(floodlightNum, LEDNum)
-			} else {
-				LED_OFF_STATE(floodlightNum, LEDNum)
-			}
-		}
-	}
-}
-#pragma GCC pop_options
-// EXTI Mains External Interrupt ISR Handler
-void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin) {
-	if(initialized) {
-	if (GPIO_Pin == Mains_Pin) {
-		mainsDetect();
-	}
-	}
-}
 void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim) {
-	if(initialized) {
-	if (htim->Instance == TIM2) {
-		pulseFloodlight();
+	if(isFloodlightRdy()) {
+		if (htim->Instance == TIM2) {
+			pulseFloodlight();
+		}
 	}
-	else if (htim->Instance == TIM3) {
+	if (htim->Instance == TIM3) {
 		//Every 1/4 of a second
 		pollKnobs();
 	}
 	else if (htim->Instance == TIM4) {
 		//Every 2ms
 		debounceButtons();
-	}
 	}
 }
 /* USER CODE END 0 */
@@ -562,9 +157,6 @@ int main(void)
 	if (initializeButtonStructs())
 		return 0;
 
-	initialized = 1;
-
-
   /* USER CODE END 2 */
 
   /* Infinite loop */
@@ -577,18 +169,8 @@ int main(void)
 		while(is_cmd_rdy())
 		{
 			light_cmd_t cmd = get_cmd();
-
-			//Send a spoof ctrl cmd back to PC
-			user_ctrl_t ctrl_input = {
-				.type = INPUT_TYPE_BUTTON,
-				.idx = 1,
-				.val = 1,
-			};
-			uart_write((char*)&ctrl_input, sizeof(user_ctrl_t));
+			updateFloodlight(cmd.lightIdx, cmd.R, cmd.G, cmd.B);
 		}
-
-		//  mainsDetect();
-		//  pulseFloodlight();
 	}
   /* USER CODE END 3 */
 }
